refactor component tree
