+++ src/pages/dashboard/document-parser/invoice.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@mui/material';
 import { useState } from 'react';
-import { DropdownSection } from 'src/components/documentParser/InvoiceParser/DropdownSection';
-import { InvoiceDetailsViewer } from 'src/components/documentParser/InvoiceParser/InvoiceDetailsViewer';
+import { DropdownSection } from 'src/components/dashboard/documentParser/InvoiceParser/DropdownSection';
+import { InvoiceDetailsViewer } from 'src/components/dashboard/documentParser/InvoiceParser/InvoiceDetailsViewer';
 import DashboardLayout from 'src/components/layouts/dashboard/DashboardLayout';
 import DashboardTemplate from 'src/components/layouts/dashboard/DashboardTemplate';
 import { parseFactura } from 'src/lib/modules/invoiceParser';
